Refactoring: actions/geo
