+++ src/actions/geo.js
@@ -15,22 +15,27 @@
  You should have received a copy of the GNU Affero General Public License
  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
-import i from 'immutable';
-import _ from 'lodash';
+export const SET_COUNTRIES = 'SET_COUNTRIES';
+export const ADD_COUNTRY = 'ADD_COUNTRY';
+export const ADD_CITY = 'ADD_CITY';
 
-import { geo } from '../actions';
+export function setCountries(countries) {
+  return {
+    type: SET_COUNTRIES,
+    countries
+  };
+}
 
-const initialState = i.Map({
-  countries: i.Map({})  // index by id
-});
+export function addCountry(country) {
+  return {
+    type: ADD_COUNTRY,
+    country
+  };
+}
 
-export default function reducer(state = initialState, action) {
-  switch (action.type) {
-    case geo.SET_COUNTRIES: {
-      state = state.set('countries', i.fromJS(_.keyBy(action.countries, 'id')));
-      break;
-    }
-  }
-
-  return state;
+export function addCity(city) {
+  return {
+    type: ADD_CITY,
+    city
+  };
 }
